🏠🐜❗ Minor Fixes ❗🐜🏠

+++ src/os/linux/install_tmux.sh
@@ -1,69 +1,69 @@
 #!/usr/bin/env bash
 
-cd "$(dirname "${BASH_SOURCE[0]}")" \
-    && . "../utils.sh"
+cd "$(dirname "${BASH_SOURCE[0]}")" &&
+  . "../utils.sh"
 
 srcdir="$(cd .. && pwd)"
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 setup_tmux() {
+  if [ -f "$HOME/.config/tmux/install.sh" ]; then
+    execute "$HOME/.config/tmux/install.sh" "Installing tmux"
+  else
+    rm -Rf ~/.tmux.conf
+    if [ -L ~/.tmux ]; then unlink ~/.tmux 2>/dev/null; fi
+    if [ -d ~/.tmux ]; then rm -Rf -f ~/.tmux 2>/dev/null; fi
 
-rm -Rf ~/.tmux.conf
-if [ -L ~/.tmux ]; then unlink ~/.tmux 2>/dev/null; fi
-if [ -d ~/.tmux ]; then rm -Rf -f ~/.tmux 2>/dev/null; fi
-
-echo ""
+    echo ""
     execute \
-        "ln -sf ~/.local/dotfiles/src/config/tmux ~/.tmux" \
-        "~/.local/dotfiles/src/config/tmux → ~/.tmux"
+      "ln -sf ~/.local/dotfiles/src/config/tmux ~/.tmux" \
+      "~/.local/dotfiles/src/config/tmux → ~/.tmux"
 
-echo ""
+    echo ""
     execute \
-        "ln -sf $srcdir/config/tmux/tmux.conf ~/.tmux.conf" \
-        "$srcdir/config/tmux/tmux.conf → ~/.tmux.conf"
+      "ln -sf $srcdir/config/tmux/tmux.conf ~/.tmux.conf" \
+      "$srcdir/config/tmux/tmux.conf → ~/.tmux.conf"
+
+  fi
 }
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 setup_tmuxplugins() {
-if [ -d ~/.local/dotfiles/src/config/tmux/tmux/plugins/tpm ]; then
+  if [ -d "$HOME/.local/share/tmux/tpm/tpm" ]; then
+    echo ""
+    execute \
+      "git -C $HOME/.local/share/tmux/tpm/tpm pull -q" \
+      "Updating tmux plugin manager"
+    echo ""
+    execute \
+      "bash $HOME/.local/share/tmux/tpm/scripts/install_plugins.sh 2> /dev/null" \
+      "Updating tmux plugins"
+  else
 
-echo ""
-     execute \
-        "cd ~/.config/tmux/plugins/tpm && \
-         git pull -q" \
-        "Updating tmux plugin manager"
-echo ""
-     execute \
-        "bash ~/.config/tmux/plugins/tpm/scripts/install_plugins.sh 2> /dev/null" \
-        "Updating tmux plugins"
-else
-
-echo ""
-     execute \
-        "rm -Rf ~/.config/tmux/plugins/tpm && \
-         git clone -q https://github.com/tmux-plugins/tpm ~/.config/tmux/plugins/tpm" \
-        "https://github.com/tmux-plugins/tpm → ~/.config/tmux/plugins/tpm"
-echo ""
-     execute \
-        "bash ~/.config/tmux/plugins/tpm/scripts/install_plugins.sh" \
-        "Installing tmux plugins → ~/.config/tmux/plugins"
-fi
+    echo ""
+    execute \
+      "rm -Rf $HOME/.local/share/tmux/tpm && \
+      git clone -q https://github.com/tmux-plugins/tpm $HOME/.local/share/tmux/tpm" \
+      "https://github.com/tmux-plugins/tpm → $HOME/.local/share/tmux/tpm"
+    echo ""
+    execute \
+      "bash $HOME/.local/share/tmux/tpm/scripts/install_plugins.sh" \
+      "Installing tmux plugins → ~/.config/tmux/plugins"
+  fi
 }
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-
 main() {
 
-    setup_tmux
+  setup_tmux
 
-    setup_tmuxplugins
+  setup_tmuxplugins
 
 }
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 main
-
